Registered a custom navigation menu.  We are using the `after_setup_theme` action hook to setup our theme.  Why are we doing setup on an action hook called 'after_setup_theme'?  Because it is after WordPress has setup the theme, not us.  This hook is the first action hook that WordPress makes available to themes and we want to do our setup the first chance we get.
We are using the `register_nav_menu()` function on line 6 to, well, register our nav menu.  The first parameter is the internal name for the theme location and the second is the visible name that WordPress will use in the admin.

If you aren't familiar with the `__()` function, it is one of the functions that WordPress uses to translate text strings.  We have to provide the text we want to translate, followed by the 'textdomain' for our theme.  You will note that the textdomain used here matches the one we set in the theme headers in the style.css file.

Now that we've added a custom nav menu, be sure to go into the WordPress admin, create a new menu, and assign it to the 'Footer Menu' theme location.