Replace the contents of the normal TwentyFifteen footer with our custom nav menu.
We are using the `wp_nav_menu()` function with a few different arguments.  First, we are setting the 'theme_location' to 'footer_menu', in order to load the menu we registered earlier here.  Next, we are setting the 'container' to 'nav'.  This will wrap our menu in a nav element instead of a div, which is a new HTML5 element and is more semantic.  Then, we set the 'container_class' to 'footer-menu'.  We do this so we will have something we can use to target our menu for styling that can't be changed from the WordPress admin.  Finally, we set the 'depth' to 1 to ensure that our footer menu doesn't display any child menu items.