Adding our own section and setting to the theme customizer.
We are using the `customize_register` action hook, which gives us a variable named `$wp_customize` that is an instance of the WP_Customize_Manager class.  We call the `add_section()` method to add our custom section.  We call the `add_setting()` method to add our custom setting.  The ID you assign the setting will be the same string you use with `get_theme_mod()` to fetch the setting's value.  Finally, we call the `add_control()` method which allows us to define what type of field (or more advanced control) will be used to control our setting.  Note that when adding a control, you must define the 'section' and 'settings' parameters using the IDs of your section and setting respectively.

Now, visit the theme customizer in the WordPress admin and check out your new setting!