Updated content.php file to display the excerpt instead of the content when on an archive page and our 'show_excerpts_in_archives' theme mod is set to true.
Note that we check for both `is_archive()` and `is_home()` because `is_home()` checks for the blog, which is technically an archive page.  If you really want to get pedantic, we could have also created a separate setting for toggling the excerpts on and off on just the blog page.

Now, go view the blog or any archive page in the theme customizer and toggle our new setting on and off to watch the magic happen!