Created a custom page template.
As long as you have a file with a 'Template Name' defined in the comment at the top of a file, this file can be used as a page template.  Just go into any page edit or creation screen in the WordPress admin and look for the 'Template' dropdown in the 'Page Attributes' metabox.  Select your new template from the dropdown and save the page.  This template will now be used when that page is loaded on the front-end.

Note that if you visit the page currently, it gives you a white screen.  This is because this template has absolutely nothing in it.  A page template gives you a completely blank slate.