Replace the content with the output of our custom query.
We are just showing the titles of the items found in a list format.  Note that we are using `get_the_title()` instead of `the_title()` because it allows you to pass in a post object or ID.  Same goes for using `get_the_permalink()` as opposed to `the_permalink()`.