Escape the title.
Escaping, as opposed to blindly outputting data, means that we clean up the output to ensure any invalid output is stripped out.  In this case, we have a page title which could potentially contain HTML.  Since an extra closing tag, or a malicious snippet, could really mess things up, we run our title through the `esc_html()` function to make sure all we display is text.

If we were able to use the `the_title()` function, WordPress would be handling escaping for us.  However, since we are using a function not intended for output, we need to do the escaping ourselves.