Escape the permalink URL.
Again, escaping is important.  Context is important when escaping.  You should always match the context in which you are outputting data to the appropriate escaping function. Since we are outputting to an HTML attribute, many people will make the mistake of using the `esc_attr()` WordPress function.  However, since we are outputting a URL, we would use `esc_url()` instead.  The `esc_url()` function is more strict than `esc_attr()`, which is why we should use it instead.

Again, if we were able to use the `the_permalink()` function, then WordPress would be handling escaping for us.  Since we are using a lower level function not intended for output, we need to do the escaping ourselves.